+++ src/Contract/Coercer.php
@@ -7,5 +7,7 @@ namespace Duon\Sire\Contract;
 /** @api */
 interface Coercer
 {
+	public string $message { get; }
+
 	public function coerce(mixed $pristine): Coercion;
 }