+++ src/Contract/Validation.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Duon\Sire\Contract;
+
+use Duon\Sire\Failure;
+
+/** @api */
+interface Validation
+{
+	public ?Failure $failure { get; }
+}