+++ src/Validator/Email.php
@@ -5,15 +5,18 @@ declare(strict_types=1);
 namespace Duon\Sire\Validator;
 
 use Duon\Sire\Contract;
+use Duon\Sire\Validation;
 use Override;
 
 /** @api */
 final class Email implements Contract\Validator
 {
-	public string $message = 'Invalid email address';
+	public string $message {
+		get => 'Invalid email address';
+	}
 
 	#[Override]
-	public function validate(Contract\Value $value, string ...$args): bool
+	public function validate(Contract\Value $value, string ...$args): Contract\Validation
 	{
 		$email = filter_var(
 			trim((string) $value->value),
@@ -23,9 +26,9 @@ final class Email implements Contract\Validator
 		if ($email !== false && ($args[0] ?? null) === 'checkdns') {
 			[, $mailDomain] = explode('@', $email);
 
-			return checkdnsrr($mailDomain, 'MX');
+			return Validation::from(checkdnsrr($mailDomain, 'MX'));
 		}
 
-		return $email !== false;
+		return Validation::from($email !== false);
 	}
 }