+++ src/Validator/Regex.php
@@ -5,15 +5,18 @@ declare(strict_types=1);
 namespace Duon\Sire\Validator;
 
 use Duon\Sire\Contract;
+use Duon\Sire\Validation;
 use Override;
 
 /** @api */
 final class Regex implements Contract\Validator
 {
-	public string $message = 'Does not match the required pattern';
+	public string $message {
+		get => 'Does not match the required pattern';
+	}
 
 	#[Override]
-	public function validate(Contract\Value $value, string ...$args): bool
+	public function validate(Contract\Value $value, string ...$args): Contract\Validation
 	{
 		// As regex patterns could contain colons ':' and validator
 		// args are separated by colons and split at their position
@@ -21,9 +24,9 @@ final class Regex implements Contract\Validator
 		$pattern = implode(':', $args);
 
 		if ($pattern === '') {
-			return false;
+			return Validation::invalid();
 		}
 
-		return preg_match($pattern, $value->value) === 1;
+		return Validation::from(preg_match($pattern, $value->value) === 1);
 	}
 }